refactor: chip state 수정 (#27)
