+++ src/shared/components/chip/chip-state/chip-state.tsx
@@ -1,0 +1,35 @@
+import { chipStateVariants } from '@components/chip/chip-state/styles/chip-state-variants';
+import { cn } from '@libs/cn';
+import type { VariantProps } from 'class-variance-authority';
+
+type ChipColorType = NonNullable<VariantProps<typeof chipStateVariants>['colorType']>;
+
+interface ChipStateProps {
+  label?: string;
+  rate?: number;
+  colorType?: ChipColorType;
+  className?: string;
+}
+
+const ChipState = ({ label, rate, colorType, className }: ChipStateProps) => {
+  if (rate !== undefined) {
+    return (
+      <div
+        data-state="rate"
+        className={cn(chipStateVariants({ colorType }), 'cap_12_m', className)}
+      >
+        <div className="flex items-center gap-[0.2rem]">
+          <span>매칭률</span>
+          <div className="flex items-center">
+            <p>{rate}</p>
+            <p>%</p>
+          </div>
+        </div>
+      </div>
+    );
+  }
+
+  return <p className={cn(chipStateVariants({ colorType }), className)}>{label}</p>;
+};
+
+export default ChipState;
