--- src/shared/components/chip/introduction.mdx
@@ -1,7 +1,0 @@
-import { Meta } from '@storybook/blocks';
-
-<Meta title="Introduction" />
-
-# MATEBALL
-
-### Design System for MATEBALL
